Restricting pages

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Room,Topic
 from .forms import RoomForm
 from django.db.models import Q
@@ -17,6 +19,9 @@ from django.contrib.auth import authenticate, login, logout
 
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method=='POST':
         username=request.POST.get('username')
         password=request.POST.get('password')
@@ -64,6 +69,7 @@ def room(request,pk):
     return render(request,'base/room.html',context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form=RoomForm()
     if request.method=='POST':
@@ -76,9 +82,13 @@ def createRoom(request):
     return render(request,'base/room_form.html',context)
 
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
 
     if request.method=='POST':
         form=RoomForm(request.POST,instance=room)
@@ -90,8 +100,13 @@ def updateRoom(request,pk):
     return render(request,'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to delete anyones\'s else room!!')
+    
     if request.method=="POST":
         room.delete()
         return redirect("home")
